New css commands for social id
Added new width commands to the social id so that when put into other languages it will be readble to users, also new media query was added for extra wide screens so the text would be readbale acorss all screen sizes. Also some tidying/formatting of the css was done to be more readable and consistent with the other css files. (the new css changes will also effect international sites be sure to check them once the update are added.)